requirements.txt
Django==5.1.8; python_version >= "3.10"
Django==4.2.17; python_version >= "3.8" and python_version < "3.10"
Django==3.2.25; python_version < "3.8"
django-environ==0.10.0
google-cloud-secret-manager==2.16.1
mysqlclient==2.1.1
djangorestframework
djangorestframework-simplejwt
